+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Install `prop-types` package in `dependencies` as official recommended.
+- Add new search and CRM icons.
+- Add new `<SearchInput>`.
 
 ### Changed
 - Upgrade `react` & `react-dom` to v15.5.